src/shop/InsertCart.ts
import Subscribe from '../utils/Subscribe';
import {MiniCartProduct} from './views/MiniCartRenderer';
import {Product} from './data/products';
import {cost_ajax} from './api/dummyProductAPI';
import {shipping_ajax} from './api/dummyBuyAPI';

export class InsertCart extends Subscribe<MiniCartProduct[]> {
  constructor(miniCartProducts: MiniCartProduct[]) {
    super(miniCartProducts);
  }

  // 함수형 코딩 394장. 타임라인 버그 재현 상태
  private total = 0;

  // 함수형 코딩 394장. 타임라인 버그 재현 메서드
  public calc_cart_total(update_total_dom: (total: number) => void) {
    const cart = [...this.miniCartProducts];

    this.total = 0;
    cost_ajax(cart, cost => {
      // console.log(`cost : ${cost}`);
      this.total = cost;
      shipping_ajax(cart, shipping => {
        // console.log(`shipping : ${shipping}`);
        this.total += shipping;
        update_total_dom(this.total);
      });
    });
  }

  public setProduct(product: Product) {
    const includeIndex = this.miniCartProducts.findIndex(
      prod => prod.name === product.name
    );

    if (includeIndex >= 0) {
      const newProducts = [...this.miniCartProducts];
      const product = newProducts[includeIndex];
      if (!product) return;
      product.quantity += 1;
      this.miniCartProducts = newProducts;
    } else {
      this.miniCartProducts = [
        ...this.miniCartProducts,
        {
          ...product,
          quantity: 1,
        },
      ];
    }

    this.next();
  }

  private get miniCartProducts() {
    return this.data;
  }

  private set miniCartProducts(miniCartProducts: MiniCartProduct[]) {
    this.data = miniCartProducts;
  }

  // api 없는 javascript 네이티브 구현
  // get totalPrice() {
  //   return this.miniCartProducts.reduce((total, product) => {
  //     total += product.price * product.quantity;
  //     return total;
  //   }, 0);
  // }
}

export default function setup(miniCartProducts: MiniCartProduct[]) {
  return new InsertCart(miniCartProducts);
}
